+++ src/components/PeopleCarts/Carts.css
@@ -1,0 +1,18 @@
+.container {
+  position: relative;
+  top: 20px;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(230px, 1fr));
+  grid-template-rows: repeat(2, 400px);
+  column-gap: 20px;
+  row-gap: 20px;
+}
+.container-small {
+  position: absolute;
+  top: 85px;
+  display: grid;
+  grid-template-columns: repeat(5, 170px);
+  grid-template-rows: repeat(2, 250px);
+  column-gap: 40px;
+  row-gap: 80px;
+}
